requirements.txt
click~=8.1.3
setuptools~=57.4.0
loguru~=0.6.0
PyYAML~=6.0